Predictable chunk sizes
This adds new padding and new fields to the format.  There were other
recent changes to the format, and together, these warrant a bump in
the format number.

This rewrites the emulator to handle this new format, which no longer
requires chunk parsing.

+++ software/player/inc/segavideo_parser.h
@@ -28,7 +28,7 @@ typedef struct SegaVideoChunkInfo {
   const uint8_t* audioStart;
   uint32_t audioSamples;
   const uint8_t* frameStart;
-  uint16_t numFrames;
+  uint32_t numFrames;
   const uint8_t* end;
   bool flipRegion;
 } SegaVideoChunkInfo;
